Separate agent.py in agent and group submodules
